requirements.txt
streamlit
plotly
numpy
scipy
pandas
matplotlib
anthropic
psycopg2-binary